Avoid producing output when checking the SSH_ASKPASS environment variable
The output confuses scp when it tries to open a session: It expects non-interactive sessions to not
have any output.

Note that the logic here works, but is incomplete: When there is no UI available we cannot use `gnome-ssh-askpass`, and instead
we should add a wrapper script here that uses a CLI fallback.